pyproject.toml
[tool.poetry]
name = "context-substrate"
version = "0.1.0"
description = "Personal Data Collector (PDC) - Context capture and retrieval infrastructure for AI agents"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
license = "MIT"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"

# API Framework
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
python-multipart = "^0.0.6"  # File upload support

# Task Queue
celery = "^5.3.4"
redis = "^5.0.1"
flower = "^2.0.1"  # Celery monitoring

# Database
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"  # PostgreSQL driver
pgvector = "^0.2.4"  # Vector similarity search

# LLM Providers
anthropic = "^0.18.0"  # Claude API
openai = "^1.10.0"  # OpenAI API (embeddings)

# Data Processing
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"

# File Processing
pypdf = "^3.17.4"  # PDF text extraction
python-docx = "^1.1.0"  # DOCX text extraction
chardet = "^5.2.0"  # Encoding detection for TXT files

# Utilities
httpx = "^0.26.0"  # Async HTTP client
tenacity = "^8.2.3"  # Retry logic
structlog = "^24.1.0"  # Structured logging
python-json-logger = "^2.0.7"

[tool.poetry.group.dev.dependencies]
# Development Tools
ruff = "^0.1.14"  # Linter and formatter
mypy = "^1.8.0"  # Type checking
ipython = "^8.20.0"  # Better REPL

[tool.poetry.group.test.dependencies]
# Testing Framework
pytest = "^8.0.0"
pytest-bdd = "^7.0.1"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
pytest-env = "^1.1.3"

# Test Utilities
httpx = "^0.26.0"  # For testing FastAPI
faker = "^22.0.0"  # Generate test data

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# ============================================================================
# Tool Configuration
# ============================================================================

[tool.ruff]
line-length = 100
target-version = "py311"

select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]

ignore = [
    "E501",  # Line too long (handled by formatter)
    "B008",  # Do not perform function calls in argument defaults
    "C901",  # Too complex (we'll handle complexity ourselves)
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

[tool.ruff.isort]
known-first-party = ["src"]

# ============================================================================
# MyPy Configuration
# ============================================================================

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "celery.*",
    "chardet.*",
    "docx.*",
    "flower.*",
    "pgvector.*",
    "pypdf.*",
    "pytest_bdd.*",
]
ignore_missing_imports = true

# ============================================================================
# Pytest Configuration
# ============================================================================

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

# BDD/Gherkin feature files
bdd_features_base_dir = "tests/features/"

# Coverage settings
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",  # Show all test outcomes
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]

markers = [
    "unit: Unit tests (fast, isolated)",
    "integration: Integration tests (slower, require services)",
    "e2e: End-to-end tests (slowest, full system)",
    "capture: Tests for capture stage",
    "enrichment: Tests for enrichment stage",
    "storage: Tests for storage stage",
    "retrieval: Tests for retrieval stage",
    "intelligence: Tests for intelligence stage",
]

# Environment variables for tests
env = [
    "ENVIRONMENT=test",
    "DATABASE_URL=postgresql://test:test@localhost:5433/pdc_test",
    "REDIS_URL=redis://localhost:6380/0",
    "LOG_LEVEL=DEBUG",
]

# ============================================================================
# Coverage Configuration
# ============================================================================

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__init__.py",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
